% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/esiprot.R
\docType{methods}
\name{esiprot,numeric-method}
\alias{esiprot}
\alias{esiprot,MassPeaks-method}
\alias{esiprot,numeric-method}
\title{ESIprot: an universal tool for charge state determination
and molecular weight calculation of proteins from electrospray ionization
mass spectrometry data calculate.}
\usage{
\S4method{esiprot}{numeric}(x, ...)
}
\arguments{
\item{x}{\code{double} consecutive peaks}

\item{zrange}{\code{double}, charge state range
\code{c(min charge, max charge)}}

\item{hydrogen}{\code{double}, hydrogen mass}
}
\description{
ESIprot: an universal tool for charge state determination
and molecular weight calculation of proteins from electrospray ionization
mass spectrometry data calculate.
}
\examples{

library("rmsi")

cytochromc <- c(651.6, 687.7, 728.1, 773.5, 825.0, 883.9, 951.7,
                1030.9, 1124.5)
esiprot(cytochromc)

p <- createMassPeaks(mass=cytochromc, intensity=rep(1, length(cytochromc)))

esiprot(p)

}
\author{
Original python implementation by Robert Winkler.

R implementation by Sebastian Gibb \email{mail@sebastiangibb.de}
}
\references{
R. Winkler. 2010.
ESIprot: a universal tool for charge state determination and molecular
weight calculation of proteins from electrospray ionization mass
spectrometry data.
Rapid Communications in Mass Spectrometry 24(3): 285-294
}

