% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gcode.R
\name{gcode}
\alias{gcode}
\title{Creates a gcode file.}
\usage{
gcode(file, x, y, resolution, wait, verbose = TRUE)
}
\arguments{
\item{file}{filename}

\item{x}{sampling area in x direction in mm}

\item{y}{sampling area in y direction in mm}

\item{resolution}{in mm}

\item{wait}{waiting time in sec}

\item{verbose}{verbose output?}
}
\description{
Creates a gcode file.
}

