% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/masterSpectrum.R
\name{masterSpectrum}
\alias{masterSpectrum}
\title{Create master spectrum}
\usage{
masterSpectrum(x, method = c("mean", "median", "sum"))
}
\arguments{
\item{x}{\code{list} of \code{\link[MALDIquant]{MassSpectrum-class}} or
\code{\link[MALDIquant]{MassPeaks-class}} objects.}

\item{method}{used aggregation function.}
}
\value{
a \code{list} of reduced \code{\link[MALDIquant]{MassSpectrum-class}}
or \code{\link[MALDIquant]{MassPeaks-class}} objects.
}
\description{
This function creates a master spectrum out of
\code{\link[MALDIquant]{MassSpectrum-class}} or
\code{\link[MALDIquant]{MassPeaks-class}} objects.
}
\examples{

library("rmsi")

s <- list(createMassSpectrum(mass=1:5, intensity=1:5),
          createMassSpectrum(mass=1:5, intensity=5:1))
masterSpectrum(s)

}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}

