% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/positioncsv.R
\name{positioncsv}
\alias{positioncsv}
\title{Creates a position csv file.}
\usage{
positioncsv(file, x, y, resolution)
}
\arguments{
\item{file}{filename}

\item{x}{sampling area in x direction in mm}

\item{y}{sampling area in y direction in mm}

\item{resolution}{in mm}
}
\description{
Creates a position csv file.
}

