% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmsi.R
\docType{package}
\name{rmsi-package}
\alias{rmsi-package}
\title{Mass Spectrometry Imaging}
\description{
This package provides some functions for Mass Spectrometry Imaging data.
}
\author{
Rober Winkler \email{robert.winkler@bioprocess.org} and
 Sebastian Gibb \email{mail@sebastiangibb.de}
}
\references{
\url{https://bitbucket.org/lababi/msi.r/}
}
\keyword{package}

