% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/coordinates.R
\name{coordinates}
\alias{coordinates}
\title{Get coordinates}
\usage{
coordinates(x)
}
\arguments{
\item{x}{\code{list} of \code{\link[MALDIquant]{MassSpectrum-class}} or
\code{\link[MALDIquant]{MassPeaks-class}} objects.}
}
\value{
a 2-column \code{matrix} (x, y) containing the coordinates.
}
\description{
This function returns the coordinates of
\code{\link[MALDIquant]{MassSpectrum-class}} or
\code{\link[MALDIquant]{MassPeaks-class}} objects.
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}

