% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topN.R
\name{topN}
\alias{topN}
\title{Find highest intensity values}
\usage{
topN(x, n = 100L)
}
\arguments{
\item{x}{\code{list} of \code{\link[MALDIquant]{MassSpectrum-class}} or
\code{\link[MALDIquant]{MassPeaks-class}} objects.}

\item{n}{\code{integer}, number of highest intensity values to preserve. If
\code{n} is larger than \code{length(x[[current]])} all intensity values
would be returned.}
}
\value{
a \code{list} of reduced \code{\link[MALDIquant]{MassSpectrum-class}}
or \code{\link[MALDIquant]{MassPeaks-class}} objects.
}
\description{
This function returns the \code{n} highest intensity value for each
\code{\link[MALDIquant]{MassSpectrum-class}} or
\code{\link[MALDIquant]{MassPeaks-class}} object.
}
\examples{

library("rmsi")

s <- createMassSpectrum(mass=1:5, intensity=1:5)
topN(list(s), n=2)
}
\author{
Sebastian Gibb \email{mail@sebastiangibb.de}
}

